cia-roblox-executor/requirements.txt
PyQt6
lupa
pyyaml
requests
plyer